backend/requirements.txt
llama-cpp-python
huggingface-hub         
mysql-connector-python  
python-dotenv           

fastapi
uvicorn
httpx
axios
cors
